+++ lib/redis_task_queue.dart
@@ -7,4 +7,5 @@ library;
 
 export 'src/client.dart';
 export 'src/task.dart' show Task, TaskContext, DeadLetter;
+export 'src/stats.dart' show QueueStats;
 export 'src/worker.dart';